feat: update slice templates for Svelte 5
