+++ packages/adapter-sveltekit/public/AlternateGrid/typescript.4.svelte
@@ -1,13 +1,8 @@
 <script lang="ts">
+	import { PrismicImage, PrismicRichText } from "@prismicio/svelte";
 	import { type Content, isFilled } from "@prismicio/client";
-	import {
-		PrismicImage,
-		PrismicRichText,
-		type SliceComponentProps,
-	} from "@prismicio/svelte";
 
-	const { slice }: SliceComponentProps<Content.PascalNameToReplaceSlice> =
-		$props();
+	export let slice: Content.PascalNameToReplaceSlice;
 </script>
 
 <section
@@ -58,7 +53,7 @@
 			</div>
 			{#if slice.primary.items && slice.primary.items.length}
 				<div class="es-alternate-grid__primary-content__items">
-					{#each slice.primary.items as item (item.title)}
+					{#each slice.primary.items as item, i}
 						<div class="es-alternate-grid__item">
 							{#if isFilled.richText(item.title)}
 								<div class="es-alternate-grid__item__heading">
