+++ packages/adapter-sveltekit/public/CustomerLogos/javascript.4.svelte
@@ -6,11 +6,10 @@
 		PrismicLink,
 	} from "@prismicio/svelte";
 
-	/* @typedef {import("@prismicio/client").Content} Content */
-	/* @typedef {import("@prismicio/svelte").SliceComponentProps} SliceComponentProps */
-
-	/* @type {SliceComponentProps<Content.${PascalNameToReplace}Slice>} */
-	const { slice } = $props();
+	/**
+	 * @type {import("@prismicio/client").Content.${PascalNameToReplace}Slice}
+	 */
+	export let slice;
 </script>
 
 <section
@@ -26,7 +25,7 @@
 		{/if}
 		{#if slice.primary.logos.length > 0}
 			<ul class="es-customer-logos__logos">
-				{#each slice.primary.logos as logo (logo.image)}
+				{#each slice.primary.logos as logo}
 					{#if isFilled.image(logo.image)}
 						<li class="es-customer-logos__logo">
 							<PrismicLink field={logo.link}>
