+++ packages/adapter-sveltekit/public/CustomerLogos/typescript.svelte
@@ -4,9 +4,11 @@
 		PrismicImage,
 		PrismicRichText,
 		PrismicLink,
+		type SliceComponentProps,
 	} from "@prismicio/svelte";
 
-	export let slice: Content.PascalNameToReplaceSlice;
+	const { slice }: SliceComponentProps<Content.PascalNameToReplaceSlice> =
+		$props();
 </script>
 
 <section
@@ -22,7 +24,7 @@
 		{/if}
 		{#if slice.primary.logos.length > 0}
 			<ul class="es-customer-logos__logos">
-				{#each slice.primary.logos as logo}
+				{#each slice.primary.logos as logo (logo.image)}
 					{#if isFilled.image(logo.image)}
 						<li class="es-customer-logos__logo">
 							<PrismicLink field={logo.link}>
