+++ packages/adapter-sveltekit/public/Hero/javascript.svelte
@@ -6,10 +6,11 @@
 	} from "@prismicio/svelte";
 	import { isFilled } from "@prismicio/client";
 
-	/**
-	 * @type {import("@prismicio/client").Content.${PascalNameToReplace}Slice}
-	 */
-	export let slice;
+	/* @typedef {import("@prismicio/client").Content} Content */
+	/* @typedef {import("@prismicio/svelte").SliceComponentProps} SliceComponentProps */
+
+	/* @type {SliceComponentProps<Content.${PascalNameToReplace}Slice>} */
+	const { slice } = $props();
 </script>
 
 <section
