+++ packages/adapter-sveltekit/public/Hero/typescript.4.svelte
@@ -1,14 +1,12 @@
 <script lang="ts">
-	import { type Content, isFilled } from "@prismicio/client";
 	import {
 		PrismicImage,
 		PrismicRichText,
 		PrismicLink,
-		type SliceComponentProps,
 	} from "@prismicio/svelte";
+	import { type Content, isFilled } from "@prismicio/client";
 
-	const { slice }: SliceComponentProps<Content.PascalNameToReplaceSlice> =
-		$props();
+	export let slice: Content.PascalNameToReplaceSlice;
 </script>
 
 <section
